+++ idea/testData/quickfix/createFromUsage/createFunction/call/extensionByExtensionReceiver/notExtensionCallable.kt
@@ -1,0 +1,12 @@
+// "Create extension function 'foo'" "false"
+// ACTION: Convert to expression body
+// ACTION: Create function 'foo'
+// ACTION: Rename reference
+// ERROR: Unresolved reference: foo
+fun bar(b: Boolean) {
+
+}
+
+fun test() {
+    bar(<caret>foo(1))
+}